+++ app/IxClient.cpp
@@ -93,8 +93,8 @@ bool IxClient::event(SDL_Event & event)
 {
     static int oldx = 0;
     static int oldy = 0;
-    static int oldRot = 0;
-    static int oldElv = 0;
+    static float oldRot = 0;
+    static float oldElv = 0;
     static float oldScl = 0;
     switch(event.type) {
         case SDL_MOUSEMOTION:
@@ -103,14 +103,14 @@ bool IxClient::event(SDL_Event & event)
                 //int h = renderer.getHeight();
                 const int x = event.motion.x;
                 const int y = event.motion.y;
-                int newRot = ((x - oldx) * 360) / renderer.getWidth();
+                float newRot = ((x - oldx) * 360.0) / renderer.getWidth();
                 if (oldy > renderer.getHeight()/2) {
                     newRot = oldRot + newRot;
                 } else {
                     newRot = oldRot - newRot;
                 }
                 while (newRot >= 360) { newRot -= 360; };
-                int newElv = ((y - oldy) * 90) / renderer.getHeight();
+                float newElv = ((y - oldy) * 90.0) / renderer.getHeight();
                 newElv = oldElv + newElv;
                 if (newElv < 0) { newElv = 0; }
                 if (newElv > 90) { newElv = 90; }
@@ -141,8 +141,8 @@ bool IxClient::event(SDL_Event & event)
         case SDL_KEYDOWN:
             switch (event.key.keysym.sym) {
                 case SDLK_0:
-                    renderer.setElevation(0);
-                    renderer.setRotation(45);
+                    renderer.setElevation(0.0);
+                    renderer.setRotation(45.0);
                     return true;
                     break;
                 case SDLK_UP:
@@ -169,11 +169,11 @@ bool IxClient::event(SDL_Event & event)
 
 void IxClient::mouse(int dx, int dy)
 {
-    int newRot = renderer.getRotation() + dx;
+    float newRot = renderer.getRotation() + dx;
     while (newRot >= 360) { newRot -= 360; };
     renderer.setRotation(newRot);
 
-    int newElv = renderer.getElevation() + dy;
+    float newElv = renderer.getElevation() + dy;
     if (newElv < -90) { newElv = -90; }
     if (newElv > 90) { newElv = 90; }
     renderer.setElevation(newElv);
